Move waitTxConfirmed from cli to web3
