+++ test/exchange.test.ts
@@ -30,9 +30,9 @@ import {
   ALPH_TOKEN_ID,
   getALPHDepositInfo,
   groupOfAddress,
-  BlockSubscription
+  BlockSubscription,
+  waitTxConfirmed
 } from '@alephium/web3'
-import { waitTxConfirmed } from '@alephium/cli'
 import { EventEmitter } from 'stream'
 import * as bip39 from 'bip39'
 import { testPrivateKey } from '@alephium/web3-test'
@@ -53,7 +53,7 @@ async function getGasFee(txIds: string[]): Promise<bigint> {
 async function waitTxsConfirmed(txIds: string[]): Promise<void> {
   const nodeProvider = web3.getCurrentNodeProvider()
   for (const txId of txIds) {
-    await waitTxConfirmed(nodeProvider, txId, 1, 1000)
+    await waitTxConfirmed(txId, 1, 1000)
   }
 }
 
@@ -214,7 +214,7 @@ class Exchange {
       throw new Error('Not enough balance')
     }
     const result = await transfer(this.wallet, user.address, ALPH_TOKEN_ID, amount)
-    await waitTxConfirmed(this.nodeProvider, result.txId, 1, 1000)
+    await waitTxConfirmed(result.txId, 1, 1000)
     this.withdrawTxs.push(result.txId)
     const remain = balance - (amount + WithdrawFee)
     if (remain === 0n) {
@@ -313,7 +313,7 @@ describe('exchange', function () {
           signerAddress: testWallet.address,
           destinations
         })
-        await waitTxConfirmed(nodeProvider, result.txId, 1, 1000)
+        await waitTxConfirmed(result.txId, 1, 1000)
         poolRewardTxNumber += 1
       }
     }
